Refactor helpers into their own module
